refactor: extract project-loop.js from project.js (PR-04)
Move all loop engine functions (18+ functions plus message handlers)
into a dedicated module. Includes loop state management, directory
watching, loop execution (coder/judge), crafting wizard, and all
loop registry message handlers. project.js reduced from 5,273 to
4,197 lines.
